+++ src/tsee/player/tsee_player_functions.h
@@ -1,6 +1,6 @@
 // Player
 
-void TSEE_Player_SetSpeed(TSEE *tsee, float speed);
-void TSEE_Player_SetJumpForce(TSEE *tsee, float jump_force);
+void TSEE_Player_SetSpeed(TSEE *tsee, double speed);
+void TSEE_Player_SetJumpForce(TSEE *tsee, double jump_force);
 void TSEE_Player_SetStepSize(TSEE *tsee, int step_size);
 void TSEE_Player_HandleInput(TSEE *tsee);
